pyproject.toml
[tool.poetry]
name = "archibald"
version = "0.1.0"
description = ""
authors = ["red <red@kalab.sk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.1"
pytesseract = "^0.3.10"
levenshtein = "^0.21.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
